refactor: "i did some mining off camera"
