+++ .dockerignore
@@ -1,0 +1,18 @@
+.git
+.venv
+__pycache__
+*.pyc
+*.pyo
+*.pyd
+.pytest_cache
+.mypy_cache
+.ruff_cache
+.tox
+.build
+build
+.dist
+dist
+*.egg-info
+.env
+.DS_Store
+.vscode
