+++ .gitignore
@@ -1,5 +1,5 @@
 .env
 .venv/
 .idea/
-src/__pycache__/
+__pycache__/
 .claude
