--- dockerfile
@@ -1,13 +1,0 @@
-FROM python:3.11-slim
-
-WORKDIR /app
-
-# Install dependencies
-RUN apt-get update && apt-get install -y gcc && rm -rf /var/lib/apt/lists/*
-
-COPY requirements.txt .
-RUN pip install -r requirements.txt
-
-COPY . .
-
-CMD ["python", "-m", "src"]
