+++ example.env
@@ -1,6 +1,5 @@
 SLACK_APP_TOKEN= # App token for your slack bot. Starts with 'xapp'
 SLACK_BOT_TOKEN= # Bot token for the bot. Starts with 'xoxb'
-SLACK_SIGNING_SECRET= # Signing secret for your slack bot
 CHANNEL_ID= # Channel ID for staff. Begins with 'C'
 SLACK_USER_TOKEN= # User token, used to delete messages of other users (Fails to delete them if you aren't an admin)
 AIRTABLE_API_KEY= # API key for airtable to keep track of threads
