+++ requirements.txt
@@ -1,0 +1,75 @@
+# This file was autogenerated by uv via the following command:
+#    uv export --format requirements.txt --no-hashes --output-file requirements.txt
+annotated-doc==0.0.4
+    # via fastapi
+annotated-types==0.7.0
+    # via pydantic
+anyio==4.12.1
+    # via
+    #   httpx
+    #   starlette
+certifi==2026.1.4
+    # via
+    #   httpcore
+    #   httpx
+    #   requests
+charset-normalizer==3.4.4
+    # via requests
+click==8.3.1
+    # via uvicorn
+colorama==0.4.6 ; sys_platform == 'win32'
+    # via click
+fastapi==0.128.8
+    # via fraudpheus
+h11==0.16.0
+    # via
+    #   httpcore
+    #   uvicorn
+httpcore==1.0.9
+    # via httpx
+httpx==0.28.1
+    # via fraudpheus
+idna==3.11
+    # via
+    #   anyio
+    #   httpx
+    #   requests
+inflection==0.5.1
+    # via pyairtable
+pyairtable==3.3.0
+    # via fraudpheus
+pydantic==2.12.5
+    # via
+    #   fastapi
+    #   pyairtable
+pydantic-core==2.41.5
+    # via pydantic
+python-dotenv==1.2.1
+    # via fraudpheus
+requests==2.32.5
+    # via pyairtable
+slack-bolt==1.27.0
+    # via fraudpheus
+slack-sdk==3.40.0
+    # via
+    #   fraudpheus
+    #   slack-bolt
+starlette==0.52.1
+    # via fastapi
+typing-extensions==4.15.0
+    # via
+    #   fastapi
+    #   pyairtable
+    #   pydantic
+    #   pydantic-core
+    #   typing-inspection
+typing-inspection==0.4.2
+    # via
+    #   fastapi
+    #   pydantic
+urllib3==2.6.3
+    # via
+    #   pyairtable
+    #   requests
+uvicorn==0.40.0
+    # via fraudpheus
