+++ src/config.py
@@ -11,7 +11,7 @@ load_dotenv()
 
 REQUIRED_ENV = (
     "SLACK_BOT_TOKEN",
-    "SLACK_SIGNING_SECRET",
+    "SLACK_APP_TOKEN",
     "CHANNEL_ID",
     "AIRTABLE_API_KEY",
     "AIRTABLE_BASE_ID",
@@ -22,11 +22,12 @@ REQUIRED_ENV = (
 env: dict[str, str] = {name: os.getenv(name, "") for name in REQUIRED_ENV}
 
 SLACK_BOT_TOKEN = env["SLACK_BOT_TOKEN"]
-SLACK_SIGNING_SECRET = env["SLACK_SIGNING_SECRET"]
-SLACK_USER_TOKEN = os.getenv("SLACK_USER_TOKEN", "")
+SLACK_APP_TOKEN = env["SLACK_APP_TOKEN"]
 CHANNEL_ID = env["CHANNEL_ID"]
 AIRTABLE_API_KEY = env["AIRTABLE_API_KEY"]
 AIRTABLE_BASE_ID = env["AIRTABLE_BASE_ID"]
+
+SLACK_USER_TOKEN = os.getenv("SLACK_USER_TOKEN", "")
 
 missing_env = [name for name, value in env.items() if not value]
 
@@ -37,10 +38,7 @@ if missing_env:
 
 # === Slack configuration ===
 
-slack_app = App(
-    token=SLACK_BOT_TOKEN,
-    signing_secret=SLACK_SIGNING_SECRET,
-)
+slack_app = App(token=SLACK_BOT_TOKEN)
 slack_client = WebClient(token=SLACK_BOT_TOKEN)
 slack_user_client = WebClient(token=SLACK_USER_TOKEN)
