+++ src/services/thread_manager.py
